CU-8694gzbud: Move CDB state code to its own module
